requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
pyyaml>=6.0
pandas>=2.0.0
scikit-learn>=1.0.0
tqdm
matplotlib
